refactor(agents): isolate before-tool call types
