+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -4,7 +4,7 @@ cbf4e2c3088f8886a7c9ea91325a66e0f0846cea21f0b2891f36399b4811306c  module/account
 6b674e7aa4006240227c16eb5c74360f1d67639e8f5580ad72499d6ed28283c4  module/account-resolution
 e5e67ddf3cab38fcbf9220bc3160715897e2709d9a9ff6ff36f1ecc9453c2367  module/agent-config-primitives
 74daa746deb548379d3f0d6eac3c4d082df1034c4360cc03bf51fee0f10a2e4d  module/agent-harness
-0a68703409f47b4aae910b5f42598badfe6ab81cf1ae96401b44651f0a30047c  module/agent-harness-runtime
+b80e26a4dcf2c246d01755b8da4bee272637f11ced1689fa561397c1979613e5  module/agent-harness-runtime
 5168648cd946abad8a92822889f13ceacc87ed502314a66190d0b1eb8ebe76ea  module/agent-media-payload
 8b406b7a50b0b4f088be869ea6a899c0d33229a5c73d3f82bda76ce0be0941b0  module/agent-runtime
 56b6d5fb6af3d95af1200065aca2e7d4f59e5fa59740505fe6ff433077ef6646  module/allow-from
